+++ lib/rspec/multiprocess_runner/command_line_options.rb
@@ -8,7 +8,7 @@ module RSpec::MultiprocessRunner
     attr_accessor :worker_count, :file_timeout_seconds, :example_timeout_seconds,
       :rspec_options, :explicit_files_or_directories, :pattern, :log_failing_files,
       :first_is_1, :use_given_order, :port, :head_node, :hostname, :max_nodes,
-      :unique_string
+      :run_identifier
 
     DEFAULT_WORKER_COUNT = 3
 
@@ -140,8 +140,8 @@ module RSpec::MultiprocessRunner
           self.max_nodes = max_nodes
         end
 
-        parser.on("-u", "--unique-string STRING", "A unique string used by nodes to confirm identity") do |string|
-          self.unique_string = string
+        parser.on("-r", "--run-identifier STRING", "A unique string used by nodes to confirm identity (e.g. a git commit hash)") do |string|
+          self.run_identifier = string
         end
 
         parser.on_tail("-h", "--help", "Prints this help") do